+++ pyproject.toml
@@ -3,17 +3,17 @@ requires = ["setuptools>=61.0.0"]
 build-backend = "setuptools.build_meta"
 
 [project]
-name = "datalab"
+name = "GOSTboundaries"
 dynamic = ["version"]
-description = "An Python exampe package from the World Bank Data Lab"
+description = "Comparison of various boundaries sources"
 readme = {file = "README.md", content-type = "text/markdown"}
 license = {file = "LICENSE.md", content-type = "text/markdown"}
-keywords = ["world bank", "template", "reproducible research"]
+keywords = ["world bank", "boundaries"]
 authors = [
-  {name="World Bank Development Data Group", email="datalab@worldbank.org"},
+  {name="Global Operations Support Team", email="gost@worldbank.org"},
 ]
 maintainers = [
-  {name="Gabriel Stefanini Vicente", email="gvicente@worldbank.org"},
+  {name="Benjamin P. Stewart", email="bstewart@worldbankgroup.org"},
 ]
 classifiers = [
     "Programming Language :: Python :: 3",
@@ -29,16 +29,19 @@ classifiers = [
 ]
 requires-python = ">=3.7"
 dependencies = [
-	"requests>=2.28.1",
-    "pycountry>=22.3.5",
+	"folium>=0.14.0",
+	"shapely>=2.0.1",
+	"rasterio>=1.3.6",
+	"pandas>=1.5.3,<2",
+	"geopandas>=0.12.2",
 ]
 [project.optional-dependencies]
 notebook = ["notebook>=6.5.2"]
 
 [project.urls]
-"Homepage" = "https://github.com/worldbank/template"
-"Bug Reports" = "https://github.com/worldbank/template/issues"
-"Source" = "https://github.com/worldbank/template"
+"Homepage" = "https://github.com/worldbank/Boundary_Comparison"
+"Bug Reports" = "https://github.com/worldbank/Boundary_Comparison/issues"
+"Source" = "https://github.com/worldbank/Boundary_Comparison"
 
 [tool.setuptools.dynamic]
-version = {attr = "datalab.__version__"}
+version = {attr = "GOSTboundaries.__version__"}
